% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nationalization-source.R
\name{nationalization}
\alias{nationalization}
\title{Nationalization}
\usage{
nationalization(
  dataSource,
  dataType,
  dataQuality,
  outputFolder,
  inequalityType
)
}
\arguments{
\item{dataSource}{Name of the file which the data are to be read from, in the form of a character string. Note that the data MUST be in CLEA format!}

\item{dataType}{Indicates the type of data file being used. Note that acceptable formats are ".csv",".xlsx",".rdata", or ".dta"}

\item{dataQuality}{Indicates the crucial columns which the user would like to use. Using "party.based" will utilize \code{pv1}/\code{pvs1} for vote totals/vote shares, which is the most
traditional option. Note that the user also has the option of "candidate.based", which utilizes \code{cv1}/\code{cvs1} if party information is missing and candidate totals/shares are desirable
as a proxy. Finally, "shares.based" is also another option to use ONLY shares, but is still in development, mostly for usage in presidential elections. Note that using "party.based"
is desirable over using "candidate.based" to be true to the nature of the party nationalization measures being generated.}

\item{outputFolder}{Character string of the location in which the user desires the output files to be written on their local machine. Note that four files will be generated at the
end of this script (i.e., party-level, national-level, constituency-level, and Gini inequality measures).}

\item{inequalityType}{This argument technically accepts any character string, but the end result is always coerced to "Gini": this function/measures/data are predicated on these
Gini inequality measures, thus it is recommended to use "Gini" because the script will override any other option anyway.}
}
\value{

}
\description{
This function creates and populates datasets of party nationalization measures,
using CLEA data with output at the national-level, party-level, and the
constituency-level.
}
