% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing-nationalization.R
\name{post.processing}
\alias{post.processing}
\title{Post-Processing Nationalization Measures}
\usage{
post.processing(folderLocation, removeSingleConstituencies, rewriteLocation)
}
\arguments{
\item{folderLocation}{Character string of the location in which nationalization data exists on the user's local machine. Note that this folder should be the location
where ALL the files should be (i.e., "C:/Users/Name/Downloads/" if the files all exist in the downloads folder).}

\item{removeSingleConstituencies}{Option for the user to remove countries that are a single, national constituency (i.e., countries/elections that are aggregated at the
national-level, where \code{cst_tot == 1}) because a number of measures hold no value for countries aggregated nationally. This argument accepts \code{TRUE} or \code{FALSE} exclusively.}

\item{rewriteLocation}{Character string of the folder location where the user would like the new, post-processed nationalization data to be rewritten. \emph{NOTE} that if
this argument is left blank/not-specified, the rewrite will take place in the \code{folderLocation} location, OVERWRITING the existing data files.}
}
\value{

}
\description{
This function serves to process completed nationalization measures to remove a number of
inappropriate values or values that are not valuable for analysis. This function mostly
cherry-picks specific cases, removing measures that do not make sense. In the future, this
script could/should evolve to include more cases/problematic measures to filter, making
this script an ever-evolving processing function.
}
