% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nationalization-source.R
\name{nationalization}
\alias{nationalization}
\title{Nationalization}
\usage{
nationalization(
  dataSource,
  dataFormat,
  outputFolder,
  inequalityType = "Gini",
  CandidateOrPartyBased = "party.based",
  filterSmallParties = FALSE
)
}
\arguments{
\item{dataSource}{Character string of the location of the CLEA data to be used for nationalization measure generation.}

\item{dataFormat}{Format of the data file being read. Acceptable formats are the following: ".csv",".xlsx",".rdata", or ".dta"}

\item{outputFolder}{Character string of the location where the user desires the output files to be written on their local machine. Note that four files will be generated at the
end of this script (i.e., party-level, national-level, constituency-level, and Gini inequality measures).}

\item{inequalityType}{Optional argument, which defaults to "Gini" to compute Gini measures of inequality. Other options are available (i.e., "RS","Atkinson","Theil","Kolm",
"var","square.var","entropy"), but "Gini" is highly recommended.}

\item{CandidateOrPartyBased}{Optional argument, which defaults to "party.based"; this argument accepts only "party.based" or "candidate.based", which allows the user to proxy
candidate votes/candidate shares for party votes/party shares. This is indended for the advanced user, and "party.based" is highly recomended for accurate measures.}

\item{filterSmallParties}{Optional argument, which defaults to \code{TRUE}. Generally, this package filters out parties that did not achieve at least five percent (\verb{5\%}) of the
national vote before computing measures of nationalization. However, the user can opt to maintain all parties in the output.}
}
\value{

}
\description{
This function creates and populates datasets of party nationalization measures,
using CLEA data with output at the national-level, party-level, and the
constituency-level.
}
\author{
Simon Page
}
